added two tests
