+++ tests/dtm/test_rgw_s3_process_restart.py
@@ -21,6 +21,7 @@
 """
 Test suite for testing RGW Process Restart with DTM enabled.
 """
+import copy
 import logging
 import multiprocessing
 import os
@@ -87,6 +88,9 @@ class TestRGWProcessRestart:
         cls.setup_type = CMN_CFG["setup_type"]
         cls.system_random = secrets.SystemRandom()
         cls.test_dir_path = os.path.join(TEST_DATA_FOLDER, "DTMTestData")
+        cls.large_workload = copy.deepcopy(HA_CFG["s3_bench_workloads"])
+        if cls.setup_type == "HW":
+            cls.large_workload.extend(HA_CFG["s3_bench_large_workloads"])
 
     def setup_method(self):
         """Setup Method"""
@@ -300,31 +304,31 @@ class TestRGWProcessRestart:
         self.log.info("STARTED: Verify continuous READ during rgw_s3 restart using pkill")
         log_file_prefix = 'test-42247'
         que = multiprocessing.Queue()
+        test_cfg = DTM_CFG["test_42247"]
 
         self.log.info("Step 1: Start write Operations :")
         self.dtm_obj.perform_write_op(bucket_prefix=self.bucket_name,
                                       object_prefix=self.object_name,
-                                      no_of_clients=self.test_cfg['clients'],
-                                      no_of_samples=self.test_cfg['samples'],
-                                      obj_size=self.test_cfg['size'],
+                                      no_of_clients=test_cfg['nclients'],
+                                      no_of_samples=test_cfg['nsamples'],
                                       log_file_prefix=log_file_prefix, queue=que)
         resp = que.get()
         assert_utils.assert_true(resp[0], resp[1])
         workload_info = resp[1]
         self.log.info("Step 2: Start READ Operations in loop in background:")
         proc_read_op = multiprocessing.Process(target=self.dtm_obj.perform_ops,
-                                               args=(workload_info, que,
-                                                     False,
-                                                     True,
-                                                     True, self.test_cfg['loop_count']))
+                                               args=(workload_info, que, False, True, True,
+                                                     self.test_cfg['loop_count']))
         proc_read_op.start()
 
-        self.log.info("Step 3 : Perform Single rgw_s3 Process Restart During Read Operations")
+        self.log.info("Step 3: Perform rgw_s3 Process Restart for %s times During Read "
+                      "Operations", DTM_CFG["rgw_restart_cnt"])
         resp = self.dtm_obj.process_restart(master_node=self.master_node_list[0],
                                             health_obj=self.health_obj,
                                             pod_prefix=const.SERVER_POD_NAME_PREFIX,
                                             container_prefix=const.RGW_CONTAINER_NAME,
-                                            process=self.rgw_process, check_proc_state=False)
+                                            process=self.rgw_process, check_proc_state=False,
+                                            restart_cnt=DTM_CFG["rgw_restart_cnt"])
         assert_utils.assert_true(resp, "Failure observed during process restart/recovery")
 
         self.log.info("Step 4: Wait for READ Operation to complete.")
@@ -332,12 +336,161 @@ class TestRGWProcessRestart:
             proc_read_op.join()
         resp = que.get()
         assert_utils.assert_true(resp[0], resp[1])
+
+        self.test_completed = True
+        self.log.info("ENDED: Verify continuous READ during rgw_s3 restart using pkill")
+
+    @pytest.mark.lc
+    @pytest.mark.dtm
+    @pytest.mark.tags("TEST-42253")
+    def test_copy_object_after_rgw_restart(self):
+        """Verify copy object after rgw restart using pkill."""
+        self.log.info("STARTED: Verify copy object after rgw restart using pkill")
+        object_name = 'object-test-42253'
+        bucket_list = list()
+        self.log.info("Step 1: Start write Operations :")
+        for i in range(0, 2):
+            bucket_name = f"bucket-test-42253-{i}"
+            resp = self.s3_test_obj.create_bucket(bucket_name=bucket_name)
+            assert_utils.assert_true(resp[0], resp[1])
+            bucket_list.append(bucket_name)
+        for size in self.test_cfg["size_list"]:
+            file_name = "{}{}".format("dtm-test-42253", size)
+            file_path = os.path.join(self.test_dir_path, file_name)
+            system_utils.create_file(file_path, size)
+            resp = self.s3_test_obj.put_object(bucket_list[0], f"{object_name}_{size}", file_path)
+            assert_utils.assert_true(resp[0], resp[1])
+        self.log.info("Step 2: Perform Single rgw Process Restart")
+        resp = self.dtm_obj.process_restart(master_node=self.master_node_list[0],
+                                            health_obj=self.health_obj,
+                                            pod_prefix=const.SERVER_POD_NAME_PREFIX,
+                                            container_prefix=const.RGW_CONTAINER_NAME,
+                                            process=self.rgw_process, check_proc_state=False)
+        assert_utils.assert_true(resp, "Failure observed during process restart")
+        self.log.info("Step 3: Perform Copy Object to bucket-2, download and verify on copied "
+                      "Objects")
+        for size in self.test_cfg["size_list"]:
+            resp = self.s3_test_obj.copy_object(source_bucket=bucket_list[0],
+                                                source_object=f"{object_name}_{size}",
+                                                dest_bucket=bucket_list[1],
+                                                dest_object=f"{object_name}_{size}")
+            assert_utils.assert_true(resp[0], resp[1])
+            file_name_copy = "{}{}".format("dtm-test-42253-copy", size)
+            file_path_copy = os.path.join(self.test_dir_path, file_name_copy)
+            resp = self.s3_test_obj.object_download(bucket_name=bucket_list[1],
+                                                    obj_name=f"{object_name}_{size}",
+                                                    file_path=file_path_copy)
+            assert_utils.assert_true(resp[0], resp[1])
+            file_name = "{}{}".format("dtm-test-42253-", size)
+            file_path = os.path.join(self.test_dir_path, file_name)
+            resp = system_utils.validate_checksum(file_path_1=file_path, file_path_2=file_path_copy)
+            assert_utils.assert_true(resp, "Checksum validation Failed.")
+        self.test_completed = True
+        self.log.info("ENDED: Verify copy object after rgw restart using pkill")
+
+    @pytest.mark.lc
+    @pytest.mark.dtm
+    @pytest.mark.tags("TEST-42254")
+    def test_copy_object_during_rgw_restart(self):
+        """Verify copy object during rgw restart using pkill."""
+        self.log.info("STARTED: Verify copy object during rgw restart using pkill")
+        object_name = 'object-test-42254'
+        workload = dict()
+        obj_list = list()
+        bucket_list = list()
+        que = multiprocessing.Queue()
+
+        self.log.info("Step 1: Start write Operations :")
+        for i in range(0, 2):
+            bucket_name = f"bucket-test-42254-{i}"
+            resp = self.s3_test_obj.create_bucket(bucket_name=bucket_name)
+            assert_utils.assert_true(resp[0], resp[1])
+            bucket_list.append(bucket_name)
+        for size in self.test_cfg["size_list"]:
+            file_name = "{}{}".format("dtm-test-42254-", size)
+            file_path = os.path.join(self.test_dir_path, file_name)
+            system_utils.create_file(file_path, size)
+            resp = self.s3_test_obj.put_object(bucket_list[0], f"{object_name}_{size}",
+                                               file_path)
+            obj_list.append(f"{object_name}_{size}")
+            assert_utils.assert_true(resp[0], resp[1])
+
+        self.log.info("Step 2: Perform Copy object to bucket-2 in background")
+        workload["source_bucket"] = bucket_list[0]
+        workload["dest_bucket"] = bucket_list[1]
+        workload["obj_list"] = obj_list
+        proc_cp_op = multiprocessing.Process(target=self.dtm_obj.perform_copy_objects,
+                                             args=(workload, que))
+        proc_cp_op.start()
+
+        self.log.info("Step 3: Perform Single m0d Process Restart")
+        resp = self.dtm_obj.process_restart(master_node=self.master_node_list[0],
+                                            health_obj=self.health_obj,
+                                            pod_prefix=const.SERVER_POD_NAME_PREFIX,
+                                            container_prefix=const.RGW_CONTAINER_NAME,
+                                            process=self.rgw_process, check_proc_state=False)
+        assert_utils.assert_true(resp, "Failure observed during process restart")
+
+        self.log.info("Step 4: Wait for copy object to finish")
+        if proc_cp_op.is_alive():
+            proc_cp_op.join()
+        resp = que.get()
+        assert_utils.assert_true(resp[0], resp[1])
+
+        self.log.info("Step 5: Perform Download and verify on copied Objects")
+        for size in self.test_cfg["size_list"]:
+            file_name_copy = "{}{}".format("dtm-test-42254-copy", size)
+            file_path_copy = os.path.join(self.test_dir_path, file_name_copy)
+            resp = self.s3_test_obj.object_download(bucket_name=bucket_list[1],
+                                                    obj_name=f"{object_name}_{size}",
+                                                    file_path=file_path_copy)
+            assert_utils.assert_true(resp[0], resp[1])
+            file_name = "{}{}".format("dtm-test-42254-", size)
+            file_path = os.path.join(self.test_dir_path, file_name)
+            resp = system_utils.validate_checksum(file_path_1=file_path, file_path_2=file_path_copy)
+            assert_utils.assert_true(resp, "Checksum validation Failed.")
+        self.test_completed = True
+        self.log.info("ENDED: Verify copy object during rgw restart using pkill.")
+
+    @pytest.mark.lc
+    @pytest.mark.dtm
+    @pytest.mark.tags("TEST-42248")
+    def test_continuous_write_during_rgw_s3_restart(self):
+        """Verify continuous WRITEs during rgw_s3 service restart using pkill."""
+        self.log.info("STARTED: Verify continuous WRITEs during rgw_s3 service restart using pkill")
+        log_file_prefix = 'test-42248'
+        que = multiprocessing.Queue()
+
+        self.log.info("Step 1: Start WRITE operation in background")
+        proc_write_op = multiprocessing.Process(target=self.dtm_obj.perform_write_op,
+                                                args=(self.bucket_name, self.object_name,
+                                                      self.test_cfg['clients'],
+                                                      self.test_cfg['samples'],
+                                                      log_file_prefix, que))
+        proc_write_op.start()
+
+        self.log.info("Step 3: Perform rgw_s3 Process Restart for %s times During Read "
+                      "Operations", DTM_CFG["rgw_restart_cnt"])
+        resp = self.dtm_obj.process_restart(master_node=self.master_node_list[0],
+                                            health_obj=self.health_obj,
+                                            pod_prefix=const.SERVER_POD_NAME_PREFIX,
+                                            container_prefix=const.RGW_CONTAINER_NAME,
+                                            process=self.rgw_process, check_proc_state=False,
+                                            restart_cnt=DTM_CFG["rgw_restart_cnt"])
+        assert_utils.assert_true(resp, "Failure observed during process restart/recovery")
+
+        self.log.info("Step 3: Wait for WRITE Operation to complete.")
+        if proc_write_op.is_alive():
+            proc_write_op.join()
+        resp = que.get()
+        assert_utils.assert_true(resp[0], resp[1])
         workload_info = resp[1]
 
-        self.log.info("Step 5: Perform Delete Operations on data written in Step 1:")
-        self.dtm_obj.perform_ops(workload_info, que, True, True, False)
+        self.log.info("Step 4: Perform READ Operation on data written in Step 1")
+        self.dtm_obj.perform_ops(workload_info, que, False, True, True)
         resp = que.get()
         assert_utils.assert_true(resp[0], resp[1])
 
         self.test_completed = True
-        self.log.info("ENDED: Verify continuous READ during rgw_s3 restart using pkill")
+
+        self.log.info("ENDED: Verify continuous WRITEs during rgw_s3 restart using pkill")
